pyproject.toml
[project]
name = "queska-backend"
version = "1.0.0"
description = "Queska - AI-Powered Premium Travel Experience Platform Backend API"
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.11,<3.13"
authors = [
    { name = "Queska Team", email = "dev@queska.com" },
]
keywords = [
    "travel",
    "booking",
    "experience",
    "fastapi",
    "mongodb",
    "ai",
    "travel-agent",
    "itinerary",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Travel",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.109.0,<0.112.0",
    "uvicorn[standard]>=0.27.0,<0.30.0",
    "pydantic>=2.6.0,<2.8.0",
    "pydantic-settings>=2.1.0,<2.3.0",
    "motor>=3.3.2,<3.5.0",
    "beanie>=1.25.0,<1.27.0",
    "python-jose[cryptography]>=3.3.0,<3.4.0",
    "passlib[bcrypt]>=1.7.4,<1.8.0",
    "redis>=5.0.0,<5.1.0",
    "celery>=5.3.0,<5.4.0",
    "httpx>=0.26.0,<0.28.0",
    "loguru>=0.7.2,<0.8.0",
    "python-multipart>=0.0.9,<0.1.0",
    "python-dotenv>=1.0.0,<1.1.0",
    "email-validator>=2.1.0,<2.2.0",
    "orjson>=3.9.0,<3.10.0",
    "aiofiles>=23.2.0,<24.0.0",
]

[project.optional-dependencies]
ai = [
    "google-generativeai>=0.4.0,<0.6.0",
    "openai>=1.12.0,<1.30.0",
    "sentence-transformers>=2.3.0,<2.4.0",
    "tiktoken>=0.6.0,<0.7.0",
]
search = [
    "elasticsearch>=8.12.0,<8.14.0",
    "meilisearch>=0.30.0,<0.32.0",
    "qdrant-client>=1.7.0,<1.9.0",
]
storage = [
    "cloudinary>=1.38.0,<1.40.0",
    "minio>=7.2.0,<7.3.0",
    "boto3>=1.34.0,<1.35.0",
    "Pillow>=10.2.0,<11.0.0",
]
notifications = [
    "sendgrid>=6.11.0,<6.12.0",
    "twilio>=8.12.0,<9.0.0",
    "firebase-admin>=6.4.0,<6.6.0",
]
payments = [
    "stripe>=8.0.0,<9.0.0",
]
geo = [
    "geopy>=2.4.0,<2.5.0",
    "shapely>=2.0.0,<2.1.0",
]
crawler = [
    "beautifulsoup4>=4.12.0,<4.13.0",
    "lxml>=5.1.0,<5.3.0",
    "scrapy>=2.11.0,<2.12.0",
    "playwright>=1.41.0,<1.45.0",
]
dev = [
    "pytest>=8.0.0,<8.2.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "pytest-cov>=4.1.0,<4.2.0",
    "black>=24.1.0,<24.4.0",
    "isort>=5.13.0,<5.14.0",
    "flake8>=7.0.0,<7.1.0",
    "mypy>=1.8.0,<1.10.0",
    "pre-commit>=3.6.0,<3.8.0",
    "faker>=22.6.0,<24.0.0",
]
all = [
    "queska-backend[ai,search,storage,notifications,payments,geo,crawler,dev]",
]

[project.urls]
Homepage = "https://queska.com"
Documentation = "https://docs.queska.com"
Repository = "https://github.com/queska/queska-backend"

[project.scripts]
queska = "app.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml", "*.html", "*.txt"]

# === Black Configuration ===
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

# === isort Configuration ===
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
known_first_party = ["app"]
known_third_party = ["fastapi", "pydantic", "beanie", "motor"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# === MyPy Configuration ===
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "beanie.*",
    "celery.*",
    "redis.*",
    "stripe.*",
    "cloudinary.*",
    "firebase_admin.*",
]
ignore_missing_imports = true

# === Pytest Configuration ===
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--asyncio-mode=auto",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# === Coverage Configuration ===
[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
show_missing = true
precision = 2

# === Ruff Linter (Alternative to flake8) ===
[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "build",
    "dist",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]
